+++ dbinfo/src/test/java/com/fsryan/forsuredb/info/TestData.java
@@ -17,16 +17,11 @@
  */
 package com.fsryan.forsuredb.info;
 
-import java.io.BufferedReader;
-import java.io.File;
-import java.io.FileReader;
-import java.io.IOException;
 import java.util.HashMap;
 import java.util.Map;
 
 public class TestData {
 
-    public static final String TEST_RES = "src/test/resources";
     public static final ColumnInfo[] DEFAULT_COLUMNS = new ColumnInfo[] {
             idCol(),
             createdCol(),
@@ -37,17 +32,6 @@ public class TestData {
     // Convenience constants
     public static final String TABLE_NAME = "test_table";
     public static final String TABLE_CLASS_NAME = "com.fsryan.test.TestTable";
-
-    public static String resourceText(String resourceName) throws IOException {
-        BufferedReader br = new BufferedReader(new FileReader(TEST_RES + File.separator + resourceName));
-        String line;
-        StringBuilder out = new StringBuilder();
-        while (null != (line = br.readLine())) {
-            out.append(line).append("\n");
-        }
-        br.close();
-        return out.toString();
-    }
 
     // Convenience methods for making data to go into the tests
     public static TableInfo.Builder table() {
