--- forsuredbapi/src/test/java/com/fsryan/forsuredb/api/TestData.java
@@ -1,40 +1,0 @@
-/*
-   forsuredbcompiler, an annotation processor and code generator for the forsuredb project
-
-   Copyright 2015 Ryan Scott
-
-   Licensed under the Apache License, Version 2.0 (the "License");
-   you may not use this file except in compliance with the License.
-   You may obtain a copy of the License at
-
-       http://www.apache.org/licenses/LICENSE-2.0
-
-   Unless required by applicable law or agreed to in writing, software
-   distributed under the License is distributed on an "AS IS" BASIS,
-   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-   See the License for the specific language governing permissions and
-   limitations under the License.
- */
-package com.fsryan.forsuredb.api;
-
-import java.io.*;
-
-public class TestData {
-
-    public static final String TEST_RES = "src/test/resources";
-
-    public static String resourceText(String resourceName) throws IOException {
-        BufferedReader br = new BufferedReader(new FileReader(TEST_RES + File.separator + resourceName));
-        String line;
-        StringBuilder out = new StringBuilder();
-        while (null != (line = br.readLine())) {
-            out.append(line).append("\n");
-        }
-        br.close();
-        return out.toString();
-    }
-
-    public static InputStream resourceStream(String resource) {
-        return TestData.class.getClassLoader().getResourceAsStream(resource);
-    }
-}
