+++ forsuredbcompiler/src/test/java/com/fsryan/forsuredb/annotationprocessor/generator/code/GetterGeneratorTest.java
@@ -1,6 +1,5 @@
 package com.fsryan.forsuredb.annotationprocessor.generator.code;
 
-import com.fsryan.forsuredb.TestData;
 import com.fsryan.forsuredb.annotationprocessor.generator.BaseGenerator;
 import com.fsryan.forsuredb.annotationprocessor.generator.GeneratorTest;
 import com.fsryan.forsuredb.info.TableInfo;
@@ -10,6 +9,9 @@ import org.junit.runners.Parameterized;
 
 import javax.tools.JavaFileObject;
 import java.util.Arrays;
+
+import static com.fsryan.forsuredb.TestData.targetTableWithChildForeignKey;
+import static com.fsryan.forsuredb.test.tools.ResourceUtil.resourceText;
 
 @RunWith(Parameterized.class)
 public class GetterGeneratorTest extends GeneratorTest<JavaFileObject> {
@@ -31,8 +33,8 @@ public class GetterGeneratorTest extends GeneratorTest<JavaFileObject> {
     public static Iterable<Object[]> data() throws Exception {
         return Arrays.asList(new Object[][] {
                 {
-                        TestData.resourceText("example_getter.txt"),
-                        TestData.targetTableWithChildForeignKey()
+                        resourceText("example_getter.txt"),
+                        targetTableWithChildForeignKey()
                 },
         });
     }
