+++ forsuredbcompiler/src/test/java/com/fsryan/forsuredb/annotationprocessor/generator/code/OrderByGeneratorTest.java
@@ -3,13 +3,15 @@ package com.fsryan.forsuredb.annotationprocessor.generator.code;
 import com.fsryan.forsuredb.annotationprocessor.generator.BaseGenerator;
 import com.fsryan.forsuredb.annotationprocessor.generator.GeneratorTest;
 import com.fsryan.forsuredb.info.TableInfo;
-import com.fsryan.forsuredb.TestData;
 import org.junit.Before;
 import org.junit.runner.RunWith;
 import org.junit.runners.Parameterized;
 
 import javax.tools.JavaFileObject;
 import java.util.Arrays;
+
+import static com.fsryan.forsuredb.TestData.targetTableWithChildForeignKey;
+import static com.fsryan.forsuredb.test.tools.ResourceUtil.resourceText;
 
 @RunWith(Parameterized.class)
 public class OrderByGeneratorTest extends GeneratorTest<JavaFileObject> {
@@ -27,8 +29,8 @@ public class OrderByGeneratorTest extends GeneratorTest<JavaFileObject> {
     public static Iterable<Object[]> data() throws Exception {
         return Arrays.asList(new Object[][] {
                 {
-                        TestData.resourceText("example_order_by.txt"),
-                        TestData.targetTableWithChildForeignKey()
+                        resourceText("example_order_by.txt"),
+                        targetTableWithChildForeignKey()
                 },
                 // TODO: tests for doc store generation
         });
