+++ forsuredbcompiler/src/test/java/com/fsryan/forsuredb/annotationprocessor/generator/code/SetterGeneratorTest.java
@@ -1,6 +1,5 @@
 package com.fsryan.forsuredb.annotationprocessor.generator.code;
 
-import com.fsryan.forsuredb.TestData;
 import com.fsryan.forsuredb.annotationprocessor.generator.BaseGenerator;
 import com.fsryan.forsuredb.annotationprocessor.generator.GeneratorTest;
 import com.fsryan.forsuredb.info.TableInfo;
@@ -10,6 +9,9 @@ import org.junit.runners.Parameterized;
 
 import javax.tools.JavaFileObject;
 import java.util.Arrays;
+
+import static com.fsryan.forsuredb.TestData.targetTableWithChildForeignKey;
+import static com.fsryan.forsuredb.test.tools.ResourceUtil.resourceText;
 
 @RunWith(Parameterized.class)
 public class SetterGeneratorTest extends GeneratorTest<JavaFileObject> {
@@ -26,8 +28,8 @@ public class SetterGeneratorTest extends GeneratorTest<JavaFileObject> {
     public static Iterable<Object[]> data() throws Exception {
         return Arrays.asList(new Object[][] {
                 {
-                        TestData.resourceText("example_relational_table_setter.txt"),
-                        TestData.targetTableWithChildForeignKey(),
+                        resourceText("example_relational_table_setter.txt"),
+                        targetTableWithChildForeignKey(),
                         SetterGeneratorTest.class
                 },
                 // TODO: test SetterGenerator.DocStore
